+++ ml_tools/path_manager/__init__.py
@@ -7,9 +7,19 @@ from ._path_tools import (
     sanitize_filename,
     list_csv_paths,
     list_files_by_extension,
-    list_subdirectories,
+    list_files_by_extension_global,
+    list_subdirectories
+)
+
+from ._path_tools_b import (
     clean_directory,
     safe_move,
+)
+
+from ._path_tools_c import (
+    get_file_hash,
+    get_size,
+    find_duplicate_files
 )
 
 from .._core import _imprimir_disponibles
@@ -21,9 +31,13 @@ __all__ = [
     "sanitize_filename",
     "list_csv_paths",
     "list_files_by_extension",
+    "list_files_by_extension_global",
     "list_subdirectories",
     "clean_directory",
     "safe_move",
+    "get_file_hash",
+    "get_size",
+    "find_duplicate_files",
 ]